Auth router & readme
